// src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          extensions?: Json
          operationName?: string
          query?: string
          variables?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      custom_pokemon: {
        Row: {
          ability: string
          created_at: string
          evs: Json | null
          gender: string | null
          id: string
          item: string | null
          ivs: Json | null
          level: number | null
          moves: string[]
          name: string
          nature: string | null
          owner_id: string
          shiny: boolean | null
          species_id: string
          stats: Json
          types: string[]
        }
        Insert: {
          ability: string
          created_at?: string
          evs?: Json | null
          gender?: string | null
          id?: string
          item?: string | null
          ivs?: Json | null
          level?: number | null
          moves: string[]
          name: string
          nature?: string | null
          owner_id: string
          shiny?: boolean | null
          species_id: string
          stats: Json
          types: string[]
        }
        Update: {
          ability?: string
          created_at?: string
          evs?: Json | null
          gender?: string | null
          id?: string
          item?: string | null
          ivs?: Json | null
          level?: number | null
          moves?: string[]
          name?: string
          nature?: string | null
          owner_id?: string
          shiny?: boolean | null
          species_id?: string
          stats?: Json
          types?: string[]
        }
        Relationships: []
      }
      lobbies: {
        Row: {
          code: string
          created_at: string
          guest_id: string | null
          host_id: string
          id: string
          seed: number[] | null
          state: string
        }
        Insert: {
          code: string
          created_at?: string
          guest_id?: string | null
          host_id: string
          id?: string
          seed?: number[] | null
          state?: string
        }
        Update: {
          code?: string
          created_at?: string
          guest_id?: string | null
          host_id?: string
          id?: string
          seed?: number[] | null
          state?: string
        }
        Relationships: []
      }
      profiles: {
        Row: {
          avatar_url: string | null
          id: string
          updated_at: string | null
          username: string | null
        }
        Insert: {
          avatar_url?: string | null
          id: string
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          avatar_url?: string | null
          id?: string
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never